Mark all *.zst files as binary
Git incorrectly infers some of the .zst files as text.  Marking them
with the binary attribute is a macro for disabling automatic end-of-line
normalization (-text), diffing (-diff), and merging (-merge).

Individuals can still override this with a custom diff driver, e.g.
one using textconv = zstdcat, if they wish using .git/info/attributes:

    *.zst diff=zstd

but as this requires a bit of individual-level config too:

    [diff "zstd"]
        textconv = zstdcat

it's not enabled by default.